+++ tosho_mplus/src/lib.rs
@@ -536,13 +536,12 @@ impl MPClient {
     /// * `writer` - The writer to write the image to.
     pub async fn stream_download(
         &self,
-        url: impl Into<String>,
+        url: impl AsRef<str>,
         mut writer: impl io::AsyncWrite + Unpin,
     ) -> ToshoResult<()> {
-        let url: String = url.into();
         let res = self
             .inner
-            .get(url)
+            .get(url.as_ref())
             .headers({
                 let mut headers = reqwest::header::HeaderMap::new();
                 headers.insert(